+++ release/src/router/e2fsprogs/lib/uuid/tst_uuid.c
@@ -144,7 +144,8 @@ main(int argc ATTR((unused)) , char **argv ATTR((unused)))
 	tv.tv_usec = 0;
 	time_reg = uuid_time(buf, &tv);
 	printf("UUID generated at %lu reports %lu (%ld.%ld)\n",
-	       time_gen, time_reg, tv.tv_sec, (long)tv.tv_usec);
+	       (unsigned long)time_gen, (unsigned long)time_reg,
+	       (long)tv.tv_sec, (long)tv.tv_usec);
 	/* allow 1s margin in case of rollover between sampling
 	 * the current time and when the UUID is generated. */
 	if (time_reg > time_gen + 1) {
